LayoutTests/svg/animations/animateTransform-hidden-no-rendering-updates.html

<!DOCTYPE HTML>
<html>

<head>
    <script src="../../resources/js-test.js"></script>
    <style>
        #container {
            width: 100px;
            height: 100px;
        }
    </style>
</head>

<body>
    <div id="container" hidden>
        <svg viewBox="0 0 100 100" width="100" height="100">
            <circle cx="50" cy="50" fill="none" r="35" stroke="currentColor"
                stroke-dasharray="164.93361431346415 56.97787143782138" stroke-width="6">
                <animateTransform attributeName="transform" type="rotate" repeatCount="indefinite" dur="1s"
                    values="0 50 50;90 50 50;180 50 50;360 50 50" keyTimes="0;0.40;0.65;1"></animateTransform>
            </circle>
        </svg>
    </div>
    <div id="console"></div>
    <script>
        window.jsTestIsAsync = true;

        let updateCount = 0;

        function sleepFor(duration) {
            return new Promise(resolve => setTimeout(resolve, duration));
        }

        async function waitForRenderingUpdateCountToSettle() {
            let renderingUpdateCount = 0;
            do {
                renderingUpdateCount = internals.renderingUpdateCount();
                // Wait slightly more than 2 frames at 60fps (~33.33ms) to ensure
                // any pending rendering updates have completed.
                await sleepFor(34);
            } while (internals.renderingUpdateCount() != renderingUpdateCount);

            internals.startTrackingRenderingUpdates();
        }

        window.addEventListener('load', async () => {
            if (!window.internals) {
                finishJSTest();
                return;
            }

            description('Test that SVG animations on hidden elements do not trigger rendering updates.');

            internals.setSpeculativeTilingDelayDisabledForTesting(true);
            
            await waitForRenderingUpdateCountToSettle();

            await sleepFor(250);

            updateCount = internals.renderingUpdateCount();

            shouldBeTrue('updateCount <= 5');

            finishJSTest();
        }, false);
    </script>
</body>

</html>
