LayoutTests/svg/animations/animateTransform-hidden-no-rendering-updates-expected.txt

Test that SVG animations on hidden elements do not trigger rendering updates.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS updateCount <= 5 is true
PASS successfullyParsed is true

TEST COMPLETE
